Reorganize packages
In order to avoid long redundant names for structs and functions, each data structure will have its own package.
With this a package, stack, has been introduced.
