+++ .travis.yml
@@ -9,8 +9,7 @@ before_install:
   - go get -t -v ./...
 
 script:
-    - go test -race -coverprofile=coverage.txt -covermode=atomic
+    - go test ./... -race -coverprofile=coverage.txt -covermode=atomic
 
 after_success:
     - bash <(curl -s https://codecov.io/bash)
-
